+++ sdk/model/signal_response.go
