--- sdk/model/tag_string.go
@@ -1,16 +1,0 @@
-/*
- * Orkes Conductor API Server
- *
- * Orkes Conductor API Server
- *
- * API version: v2
- * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
- */
-
-package model
-
-type TagString struct {
-	Key   string `json:"key,omitempty"`
-	Type_ string `json:"type,omitempty"`
-	Value string `json:"value,omitempty"`
-}
