+++ sdk/model/workflow_schedule_execution_model.go
@@ -1,15 +1,6 @@
-// Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
-// the License. You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
-// an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
-// specific language governing permissions and limitations under the License.
-
 package model
 
-type WorkflowScheduleExecutionModel struct {
+type WorkflowScheduleExecution struct {
 	ExecutionId          string                `json:"executionId,omitempty"`
 	ExecutionTime        int64                 `json:"executionTime,omitempty"`
 	Reason               string                `json:"reason,omitempty"`
@@ -20,5 +11,4 @@ type WorkflowScheduleExecutionModel struct {
 	State                string                `json:"state,omitempty"`
 	WorkflowId           string                `json:"workflowId,omitempty"`
 	WorkflowName         string                `json:"workflowName,omitempty"`
-	ZoneId               string                `json:"zoneId,omitempty"`
 }
